# lms-server/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: LMS API
  description: |
    Multi-tenant Learning Management System API
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Multi-Tenancy
    All resources are scoped by tenant. The tenant ID is automatically extracted from:
    - `x-tenant-id` header, or
    - The authenticated user's tenant
    
    ## Rate Limiting
    - General API: 100 requests per 15 minutes
    - Auth endpoints: 5 requests per 15 minutes
    - Password reset: 3 requests per hour
    - Tenant signup: 2 requests per hour
  version: 1.0.0
  contact:
    name: API Support
    email: edinmavric10@gmail.com
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server (example)
...
paths:
...
  /courses:
    get:
      tags:
        - Courses
      summary: Get all courses
      operationId: getAllCourses
      security:
        - bearerAuth: []
      parameters:
        - name: professor
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Courses
      summary: Create course
      description: Create new course (admin or professor)
      operationId: createCourse
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - professor
              properties:
                name:
                  type: string
                description:
                  type: string
                professor:
                  type: string
                  description: User ID with professor role
                students:
                  type: array
                  items:
                    type: string
                price:
                  type: number
                schedule:
                  $ref: '#/components/schemas/CourseSchedule'
      responses:
        '201':
          description: Course created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /courses/{id}:
    get:
      tags:
        - Courses
      summary: Get course by ID
      operationId: getCourseById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: Course details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Courses
      summary: Update course
      operationId: updateCourse
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                professor:
                  type: string
                students:
                  type: array
                  items:
                    type: string
                price:
                  type: number
                schedule:
                  $ref: '#/components/schemas/CourseSchedule'
      responses:
        '200':
          description: Course updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Courses
      summary: Delete course (soft delete)
      operationId: deleteCourse
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: Course deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /attendance:
    get:
      tags:
        - Attendance
      summary: Get all attendance records
      operationId: getAllAttendance
      security:
        - bearerAuth: []
      parameters:
        - name: student
          in: query
          schema:
            type: string
          description: Filter by student ID
        - name: lesson
          in: query
          schema:
            type: string
          description: Filter by lesson ID
        - name: status
          in: query
          schema:
            type: string
            enum: [present, absent, late, excused]
          description: Filter by attendance status
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: Filter by start date (inclusive)
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: Filter by end date (inclusive)
      responses:
        '200':
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Attendance'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Attendance
      summary: Create attendance record
      description: Record attendance for a specific lesson (admin or course professor only, prevents duplicates per student+lesson)
      operationId: createAttendance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - student
                - lesson
                - date
                - status
              properties:
                student:
                  type: string
                  description: Student ID
                lesson:
                  type: string
                  description: Lesson ID (required)
                date:
                  type: string
                  format: date
                  description: Attendance date
                status:
                  type: string
                  enum: [present, absent, late, excused]
                  description: Attendance status
      responses:
        '201':
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /attendance/{id}:
    get:
      tags:
        - Attendance
      summary: Get attendance by ID
      operationId: getAttendanceById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AttendanceId'
      responses:
        '200':
          description: Attendance details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Attendance
      summary: Update attendance
      operationId: updateAttendance
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AttendanceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [present, absent, late, excused]
                date:
                  type: string
                  format: date
      responses:
        '200':
          description: Attendance updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Attendance
      summary: Delete attendance
      operationId: deleteAttendance
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AttendanceId'
      responses:
        '200':
          description: Attendance deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  parameters:
...
    CourseId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Course ID
...
    AttendanceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Attendance ID
...
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    User:
      type: object
      properties:
        _id:
          type: string
        tenant:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, professor, student]
        status:
          type: string
          enum: [active, pending, disabled]
        pendingApproval:
          type: boolean
        approvedBy:
          type: string
        approvedAt:
          type: string
          format: date-time
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Course:
      type: object
      properties:
        _id:
          type: string
        tenant:
          type: string
        name:
          type: string
        description:
          type: string
        professor:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
        students:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: '#/components/schemas/User'
        price:
          type: number
        schedule:
          $ref: '#/components/schemas/CourseSchedule'
        isDeleted:
          type: boolean
        createdBy:
          type: string
        updatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
    CourseSchedule:
      type: object
      properties:
        days:
          type: array
          items:
            type: string
        startTime:
          type: string
        endTime:
          type: string
...
    Attendance:
      type: object
      properties:
        _id:
          type: string
        tenant:
          type: string
        student:
          oneOf:
            - type: string
            - type: object
              properties:
                _id:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
        lesson:
          oneOf:
            - type: string
            - type: object
              properties:
                _id:
                  type: string
                title:
                  type: string
                date:
                  type: string
                  format: date
                startTime:
                  type: string
                endTime:
                  type: string
                course:
                  oneOf:
                    - type: string
                    - type: object
                      properties:
                        _id:
                          type: string
                        name:
                          type: string
        date:
          type: string
          format: date
          description: Attendance date
        status:
          type: string
          enum: [present, absent, late, excused]
          description: Attendance status
        recordedBy:
          oneOf:
            - type: string
            - type: object
              properties:
                _id:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
        recordedAt:
          type: string
          format: date-time
          description: When the attendance was recorded
        createdBy:
          type: string
        updatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'